Split up block-generator commands by their package.
